pyproject.toml
[tool.poetry]
name = "beuchat-reception"
version = "0.1.0"
description = ""
authors = ["PhiC Perso <phic13@orange.fr>"]
readme = "README.md"
packages = [
    { include = "beuchat_reception", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.14"
fastapi = ">=0.133.0,<0.134.0"
uvicorn = ">=0.41.0,<0.42.0"
mysql-connector-python = "*"
requests = "^2.32.5"
python-dotenv = "^1.2.1"
bs4 = "^0.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
